refactor: 설정 항목 아이템을 별도의 파일로 옮김
