+++ feature/settings/src/main/java/com/practice/settings/Settings.kt
@@ -1,13 +1,9 @@
 package com.practice.settings
 
-import android.widget.Toast
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxSize
@@ -17,7 +13,6 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Switch
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
@@ -28,19 +23,16 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.semantics.clearAndSetSemantics
-import androidx.compose.ui.semantics.contentDescription
-import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.practice.designsystem.LightPreview
 import com.practice.designsystem.components.BlindarTopAppBar
-import com.practice.designsystem.components.LabelMedium
 import com.practice.designsystem.components.TitleMedium
 import com.practice.designsystem.theme.BlindarTheme
 import com.practice.preferences.preferences.MainScreenMode
+import com.practice.settings.items.SetDailyAlarmItem
+import com.practice.settings.items.SetDailyModeItem
 import com.practice.settings.uistate.SettingsUiState
 
 @Composable
@@ -146,121 +138,13 @@ private fun SettingsItems(
     modifier: Modifier = Modifier,
 ) {
     Column(modifier = modifier) {
-        DailyModeSetting(
+        SetDailyModeItem(
             isDailyModeEnabled = uiState.mainScreenMode == MainScreenMode.Daily,
             onToggle = onToggleDailyMode,
         )
-        DailyAlarmSetting(
+        SetDailyAlarmItem(
             isDailyAlarmEnabled = uiState.isDailyAlarmEnabled,
             onToggle = onToggleDailyAlarm,
-        )
-    }
-}
-
-@Composable
-private fun DailyModeSetting(
-    isDailyModeEnabled: Boolean,
-    onToggle: (Boolean) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val titleDescription =
-        stringResource(id = if (isDailyModeEnabled) R.string.settings_daily_mode_title_description_enabled else R.string.settings_daily_mode_title_description_disabled)
-    val onClickLabel =
-        stringResource(id = if (isDailyModeEnabled) R.string.settings_daily_mode_disable else R.string.settings_daily_mode_enable)
-    val onToggleMessage =
-        stringResource(id = if (isDailyModeEnabled) R.string.settings_daily_mode_disabled else R.string.settings_daily_mode_enabled)
-
-    val context = LocalContext.current
-    val afterToggle = {
-        Toast.makeText(context, onToggleMessage, Toast.LENGTH_SHORT).show()
-    }
-
-    Row(
-        modifier = modifier
-            .semantics(mergeDescendants = true) {}
-            .clickable(
-                onClick = {
-                    onToggle(!isDailyModeEnabled)
-                    afterToggle()
-                },
-                onClickLabel = onClickLabel,
-            )
-            .padding(16.dp),
-        horizontalArrangement = Arrangement.spacedBy(8.dp),
-    ) {
-        Column(
-            modifier = Modifier
-                .weight(1f)
-                .padding(end = 16.dp),
-            verticalArrangement = Arrangement.spacedBy(8.dp),
-        ) {
-            TitleMedium(
-                text = stringResource(id = R.string.settings_daily_mode_title),
-                modifier = Modifier.clearAndSetSemantics {
-                    contentDescription = titleDescription
-                },
-            )
-            LabelMedium(text = stringResource(id = R.string.settings_daily_mode_body))
-        }
-        Switch(
-            checked = isDailyModeEnabled,
-            onCheckedChange = {
-                onToggle(!isDailyModeEnabled)
-                afterToggle()
-            },
-            modifier = Modifier.clearAndSetSemantics { },
-        )
-    }
-}
-
-@Composable
-private fun DailyAlarmSetting(
-    isDailyAlarmEnabled: Boolean,
-    onToggle: (Boolean) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val permissionLauncher = rememberNotificationPermissionLauncher(isDailyAlarmEnabled, onToggle)
-
-    val context = LocalContext.current
-    val onClick = {
-        onDailyAlarmClick(!isDailyAlarmEnabled, context, permissionLauncher) {
-            onToggle(!isDailyAlarmEnabled)
-        }
-    }
-
-    val titleDescription =
-        stringResource(id = if (isDailyAlarmEnabled) R.string.settings_daily_alarm_title_description_enabled else R.string.settings_daily_alarm_title_description_disabled)
-    val onClickLabel =
-        stringResource(id = if (isDailyAlarmEnabled) R.string.settings_daily_alarm_disable else R.string.settings_daily_alarm_enable)
-
-    Row(
-        modifier = modifier
-            .semantics(mergeDescendants = true) {}
-            .clickable(
-                onClick = onClick,
-                onClickLabel = onClickLabel,
-            )
-            .padding(16.dp),
-        horizontalArrangement = Arrangement.spacedBy(8.dp),
-    ) {
-        Column(
-            modifier = Modifier
-                .weight(1f)
-                .padding(end = 16.dp),
-            verticalArrangement = Arrangement.spacedBy(8.dp),
-        ) {
-            TitleMedium(
-                text = stringResource(id = R.string.settings_daily_alarm_title),
-                modifier = Modifier.clearAndSetSemantics {
-                    contentDescription = titleDescription
-                },
-            )
-            LabelMedium(text = stringResource(id = R.string.settings_daily_alarm_body))
-        }
-        Switch(
-            checked = isDailyAlarmEnabled,
-            onCheckedChange = { onClick() },
-            modifier = Modifier.clearAndSetSemantics { },
         )
     }
 }
